requirements.txt
streamlit
SpeechRecognition
ipykernel
pyaudio
sounddevice
noisereduce
torch
google.generativeai
vertexai
Vertex